// lib/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      projects: {
        Row: {
          created_at: string
          creator_id: string
          description: string | null
          id: string
          name: string
          project_covers: string | null
          updated_at: string
        }
        Insert: {
          created_at?: string
          creator_id: string
          description?: string | null
          id?: string
          name: string
          project_covers?: string | null
          updated_at?: string
        }
        Update: {
          created_at?: string
          creator_id?: string
          description?: string | null
          id?: string
          name?: string
          project_covers?: string | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "projects_creator_id_fkey"
            columns: ["creator_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
    Enums: {
      notification_status: "UNREAD" | "READ"
      notification_type:
        | "PROJECT_INVITATION"
        | "TASK_ASSIGNED"
        | "TASK_DUE_SOON"
        | "TASK_OVERDUE"
      roles: "VIEWER" | "MEMBER" | "ADMIN" | "OWNER"
      task_priority: "LOW" | "MEDIUM" | "HIGH"
      task_status: "BACKLOG" | "IN_PROGRESS" | "COMPLETED"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never